requirements.txt
torch
torchvision
timm
numpy
pandas
matplotlib
scikit-learn
opencv-python
Pillow
tqdm
